Add enum for local storage and refactor data retrieve from api calls
