+++ src/pages/index.tsx
@@ -1,2 +1,1 @@
 export * from "./landing";
-export * from "./pages";
