+++ src/pages/landing/index.tsx
@@ -10,7 +10,7 @@ import { Icon, Navbar, ContactModal } from "@components/common";
 import Phone from "@assets/images/i6.png";
 import Play from "@assets/images/play.png";
 import { urlFormat } from "@/utils";
-import { Pages } from "../pages";
+import { Pages } from "@/types";
 
 import styles from "./Landing.module.css";
