+++ src/types/index.tsx
@@ -1,18 +1,2 @@
-export type UserType = {
-  "access-token": string;
-  avatar: object;
-  client: string;
-  email: string;
-  expiry: number | null;
-  first_name: string;
-  gender: string;
-  id: number;
-  last_name: string;
-  name: string;
-  uid: string;
-  username: string;
-};
-
-export type GenderType = {
-  option: string;
-};
+export * from "./types";
+export * from "./enums";
